+++ pages/teacher/ajax.php
@@ -2,14 +2,14 @@
 if (isset($_SESSION['user_id'])) {
     $mode = $_POST['mode'];
 
-    $fname = $_POST['fname'];
-    $lname = $_POST['lname'];
-    $phone = $_POST['phone'];
-    $email = $_POST['email'];
-    $at = $_POST['at'];
-    $pass = $_POST['phone'];
-    $user_role = $_POST['user_role'];
-    $tuluv = $_POST['tuluv'];
+    $fname = @$_POST['fname'];
+    $lname = @$_POST['lname'];
+    $phone = @$_POST['phone'];
+    $email = @$_POST['email'];
+    $at = @$_POST['at'];
+    $pass = @$_POST['phone'];
+    $user_role = @$_POST['user_role'];
+    $tuluv = @$_POST['tuluv'];
     if ($mode == 1) {
         $id = $_POST['id'];
         $success = _exec(
@@ -29,13 +29,29 @@ if (isset($_SESSION['user_id'])) {
         echo "Амжилттай!";
     } elseif ($mode == 3) {
         $id = $_POST['id'];
-        $success = _exec(
-            "DELETE FROM teacher WHERE id = ?",
-            'i',
-            [$id],
-            $count
+        _selectRowNoParam(
+            "SELECT COUNT(id) FROM `class` WHERE teacherid = $id",
+            $too
         );
-        echo "Амжилттай!";
+
+        _selectRowNoParam(
+            "SELECT COUNT(id) FROM `att` WHERE tid = $id",
+            $atoo
+        );
+
+        if ($too > 0) {
+            echo "Багш даасан ангитай тул устгах боломжгүй!";
+        } else if ($atoo > 0) {
+            echo "Багш ирц бүртгэсэн тул устгах боломжгүй!";
+        } else {
+            $success = _exec(
+                "DELETE FROM teacher WHERE id = ?",
+                'i',
+                [$id],
+                $count
+            );
+            echo "Амжилттай!";
+        }
     }
 ?>
 <?php
